[bugfix, bugzilla 1107289] Compilation problems with JDK1.5 and Tomcat 5.5 issue
The `bug' is caused again by Saxon and has nothing to do with JDK. Saxon overrides the default XML parser with a non-validating Aelfred
and causes Tomcat to crash. The only workaround I found for this is to physically remove 'services' folder from Saxon's JAR and that's basically what I did.
Saxon is used only in xslt-output renderer and only because we have a custom defined Java function in there... One day I'll rewrite it to plain XSLT just for the sake
of portability.


git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk/carrot2@468 7ff1d41c-760d-0410-a7ff-a3a56f310b35