Minor fixes to the build process: ANT would fail if invoked like this:
ant -Dcopy.dependencies=true -q build

on a fresh build. Reason: bootstrap JAR was created successfully first, then replaced with another JAR created by the proper build process. Because ANT already locked the JAR file and read task definitions from there, it somehow used fixed ZIP offsets to other classes and failed to find them when the JAR was replaced... if this explanation is vague that's because it is vague to me. Anyway, it builds now without problems.


git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk/carrot2@478 7ff1d41c-760d-0410-a7ff-a3a56f310b35