A bug fix.
When Multilingual strategy is used and the language is not specified in the input there are no stopwords
in the features vector. The counting code previously relied on the presence of a stopword that would separate
terms from phrases. The new code doesn't make this assumption and seems to work fine.


git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@584 7ff1d41c-760d-0410-a7ff-a3a56f310b35