- Bugfix: Long urls and text sequences on the clusters pane. https://sourceforge.net/tracker/index.php?func=detail&aid=1422480&group_id=85379&atid=576012
Fixed to the extent possible, the following changes have been introduced:
a) Long URLs should break at "/;.-?" characters.
b) the cluster display pane is not of fixed size. It can stretch from the minimum (~250 pixels) to a wider size. Therefore you will see the cluster label entirely
and it won't go over the results. The downside is that you'll have a scrollbar on the entire page.

Breaking long text sequences is more trouble then gain, so I leave it as is. We should implement a frame-based output component one day and this problem will not be an issue then.

- Placed a non-breakable space between the last word of the cluster's label and the number of its documents.


git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk/carrot2@954 7ff1d41c-760d-0410-a7ff-a3a56f310b35