http://issues.carrot2.org/browse/CARROT-144
- Rewritten XmlLocalInputComponent. Default XSLTs are precompiled. 
- Cleanups in the architecture between XmlLocalInputComponent and subclasses.

This is a major update that can result in different runtime behavior. One specific scenario is this: when a subclass passed
a source URL as a string with substitutable parameters, it expected the default URL to be re-evaluated for each query. This
is no longer the case (default parameters are evaluated early). A clean fix for this is to pass a null source to the 
superclass (XmlLocalInputComponent) and force source argument on each query.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@2174 7ff1d41c-760d-0410-a7ff-a3a56f310b35