http://issues.carrot2.org/browse/CARROT-145
One possible fix for the above bug. It's not an ideal solution though as it requires either an explicit call to LocalControllerBase#cleanup() or garbage collection of the LocalControllerBase instance (finalize is overriden).

I am committing this only to keep the patch in SVN history, will revert it and rewrite the code on top of a simpler pool (KeyedStack...).

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk/carrot2@2178 7ff1d41c-760d-0410-a7ff-a3a56f310b35