carrot2-core: Ula's patch fixed -- we do want all exceptions to propagate, hence the catch clause removed, the loop will be broken anyway. Updated a passing test that should fail because of the bug. Also updated two tests that failed after Ula's patch. Ula: if you change any of our code, please run tests to make sure everything is still working or update the tests to reflect the conceptual changes made to the code. Incidentally, maybe it's time to set up some provisional build for this branch, so that tests get executed automatically?
carrot2-util-attribute: added the "required" parameter to the @Input annotation. Adding a @Required annotation instead is tempting, but then it would seem it also applies to @Output attributes, which is not true. On the other hand, constraints also apply to @Input attributes, but they still have separate annotations. I need to think this through...

carrot2-source-yahoo: I made a horrible mistake of applying the Clean Up wizard to the code, which made some of the fields final. As a result, the default values of attributes would never get overriden. I hope I reverted the problems across all the projects.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/branches/3.0-proto@2231 7ff1d41c-760d-0410-a7ff-a3a56f310b35