CARROT-279: eTools character encoding issues.
Fixed the broken platform-dependent InputStreamReader occurrence. In general, if you use InputStreamReader or OutputStreamWriter, you should ALWAYS make sure to pass the encoding parameter and never rely on the platform's specific. I know everyone knows that, but somehow this slipped through...

Is there a FindBugs check for this? There certainly should be...

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/branches/3.0-proto@2725 7ff1d41c-760d-0410-a7ff-a3a56f310b35