http://issues.carrot2.org/browse/CARROT-301
Added Google Appliance input component based on the XML format returned by Google Appliance. The current limitation is that only up to 100 results can be fetched from the source (and this cannot be easily avoided, unless we make major modifications to the xml-input engine, which I do not think make sense in the 2.x line anymore.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk/carrot2@2800 7ff1d41c-760d-0410-a7ff-a3a56f310b35