http://issues.carrot2.org/browse/CARROT-299
Include this for the M4 as well (and don't close it yet).

This patch removes previous separate extension points for algorithms and sources, replacing them with an extension point that gives a resource path to a component suite. This way component suite definition can be used to drive Workbench sources and algorithms. Notes:

- Something is still not working properly with initialization parameters because Yahoo and Yahoo News return the same result.
- Remember to add 'carrot2-component-suites' project to the launching config. This project contains workbench-suite.xml file with Workbench sources and algorithms.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/branches/3.0-proto@2822 7ff1d41c-760d-0410-a7ff-a3a56f310b35