http://issues.carrot2.org/browse/CARROT-385
Refactoring of the ValueHintEnum constraint and its associated editors. I checked on various examples and with various components -- everything seems all right. The only thing I'm not 100% sure is if the editor should write back (to the attribute set) a String value or an Enum constant. The binder takes care of the conversion to enums and this solution is quite nice with regard to  String attributes with enum hints, so I leave it for now.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/branches/3.0-proto@2986 7ff1d41c-760d-0410-a7ff-a3a56f310b35