http://issues.carrot2.org/browse/CARROT-418
Right, maybe having thread-safe language models was a little over the top here. I changed the implementation to not-thread-safe, although linguistic resources are still cached and reused (if possible). All the tests pass and I think the implementation is a bit cleaner as well. Apologies for the optimized imports, but it somehow happened and I had no time to revert file by file.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/branches/3.0-proto@3109 7ff1d41c-760d-0410-a7ff-a3a56f310b35