http://issues.carrot2.org/browse/CARROT-463 http://issues.carrot2.org/browse/CARROT-461
I eventually decided to leave a single class for all XML input (although I am still hesitating about this). Added a boolean attribute to read all documents from the input. Added a named attribute to force the title of visual elements (tab, top banner) in the Workbench. 

A small issue remains: initially the title is the input query and it is later being replaced with the actual query taken from the XML (or the name of the resource). This will only occur with the XML input source, so it's better evil than if we picked an empty initial title (which would cause a similar flicker for all other components).

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@3310 7ff1d41c-760d-0410-a7ff-a3a56f310b35