http://issues.carrot2.org/browse/CARROT-535
A fix for wrong document counts in internal nodes of the GST that could manifest itself with certain input. 
This is not the optimal algorithmic solution (adds some overhead), but working out how to update the GST
(in this old implementation) is not trivial at all because sentence termination markers are not unique 
(so they can end up in implicit or leaf nodes). This solution should work and I still think we should switch 
to using suffix arrays at some point in the future.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@3558 7ff1d41c-760d-0410-a7ff-a3a56f310b35