carrot2-core: After some thought, I think fiddling with the way @Output attributes are collected (treat one special value (null) in a different way and collect it or not) would bring more harm than good. After all, if components want to cooperate through a common attribute, it's best to make this communication explicit. In the specific case of active-language passing, this would mean all components that declare the this attribute should use a common default value (defined on one of the components, DefaultLanguageModelFactory in that specific case). I'm leaving the tests as the documentation of the consequences of sharing @Input @Output attributes between components.
Incidentally, @Init @Output attributes are collected during initialization to the attribute map in SimpleController, this may be the reason for strange test results. Now when I think about it, I'm not sure if this was the best idea... The current implementation makes Simple and Caching controller not fully interchangeable (no access to @Init @Output attributes in CachingController). I'll need to think how to solve it best.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@3587 7ff1d41c-760d-0410-a7ff-a3a56f310b35