http://issues.carrot2.org/browse/CARROT-543
Added an example of using DCS to accomplish simple faceting by a custom document field. The example is for CURL only, because it is really about passing a different document field name to the "by source" algorithm inside the DCS. There are existing examples of using custom attributes for all other languages that apply further on.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@3604 7ff1d41c-760d-0410-a7ff-a3a56f310b35