http://issues.carrot2.org/browse/CARROT-595
A closer look at the problem revealed that #parse directive repeatedly reads the resource off disk and parses it in the tight loop. Replaced the code with a velocity macro and it speeds up a great deal. See if we can close this issue -- to me, the speed now is acceptable, I left the limits of 1000 documents for decency.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@3839 7ff1d41c-760d-0410-a7ff-a3a56f310b35