http://issues.carrot2.org/browse/CARROT-609
Added initial support for an annotation processor to generate @Bindable metadata. It will be
so much cleaner and powerful compared to the ANT-based version we currently have. qdox will
still be used because I don't want to use a custom JavaDoc parser, but we will use qdox for parsing
JavaDoc only, the rest of the data will be acquired from javac provided data.

I firmly believe this should make it for the next release of Carrot2, it will make the build
infrastructure easier to handle. I will work on this during the weekend, time permitting.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@4414 7ff1d41c-760d-0410-a7ff-a3a56f310b35