http://issues.carrot2.org/browse/CARROT-307
Welcome view is a singleton now. Added navigation bar so that when links are clicked the user may
click back. A better alternative would be to open links in an external browser (or another editor),
but this would require creating a custom editor (with embedded Browser) because IWebBrowser doesn't
provide any means of customization.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@4457 7ff1d41c-760d-0410-a7ff-a3a56f310b35