http://issues.carrot2.org/browse/CARROT-736 http://issues.carrot2.org/browse/LINGO-210
Fix for the AIOOB bug. PoolingProcessingComponentManager's reset listener was using system identity
hash code of the currently processing component to place its attributes in the map for later resetting.
When two components had the same system identity (and yes, it turned out to be possible -- had this
before my own eyes in the debug view of Eclipse) there was a possibility that ANOTHER component would
be reset with the attributes of the previous component, leading to reuse of component fields. This of
course was causing various problems with non-thread safe classes, such as stemmers.

The fix in this patch does two things:

- it removes one ConcurrentHashMap that was used solely for keeping the link between a component and its
ID (I do understand the clean API rationale, but this is an extra overhead I don't find an excuse for --
the prepare method accepts an ID, so it's symmetric anyway).

- reset listener now uses a ReferenceEquality decorator to store the component's attributes
in the concurrent hash map. I don't know if the concurrent map is inevitable here, but didn't
investigate any further.

git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@4610 7ff1d41c-760d-0410-a7ff-a3a56f310b35