I'm fixing the tests, but not the problem...
Short synopsis: the problem here is that SimpleXmlWrappers has a static
global facility for registering wrappers. This global facility is cleaned
during tests. The test that was failing was ensuring that registration has
been executed by loading the class that does it in a static initializer:

        // Just load LuceneDocumentSource to install the wrapper
        ReflectionUtils.classForName(LuceneDocumentSource.class.getName());

this however does nothing if the class is already loaded. In a situation
when tests were interleaved like this:

LuceneDocumentSourceTest (registers wrappers)
SimpleXmlWrappersTest (clears wrappers globally)
FSDirectoryWrapperTest (attempts to load a class and register wrappers -
nop, then fails because wrappers are empty).

the execution ended in a failure. All other arrangements of tests ended up
successfully. The question is how to fix the above? A few options:

1) are these wrappers still necessary? any improvements in simplexml?
2) can we make simplexmlwrappers non-static?